Refactored BatteryController & added unit tests for it

+++ src/batterycontroller.h
@@ -11,11 +11,11 @@ class BatteryController : public BatterySensor
     static constexpr const float kStepsToFullAmount = 20.0f;
     static constexpr const float kDischargeUnit = 1.0f;
 
-    const unsigned int full_amount;
+    const float full_amount;
     float current_amount;
 
 public:
-    BatteryController(unsigned int full_amount) : full_amount(full_amount), current_amount(full_amount) {}
+    BatteryController(unsigned int full_amount) : full_amount((float)full_amount), current_amount(full_amount) {}
 
     virtual float getCurrentAmount() const
     {
@@ -24,8 +24,8 @@ public:
 
     virtual void charge()
     {
-        float updated_amount = current_amount + (float)full_amount / kStepsToFullAmount;
-        current_amount = std::min(updated_amount, (float)full_amount);
+        float updated_amount = current_amount + full_amount / kStepsToFullAmount;
+        current_amount = std::min(updated_amount, full_amount);
     }
 
     virtual void discharge()
